pyproject.toml
[project]
name = "clinvar_ingest"
authors = [
    {name = "Kyle Ferriter", email = "kferrite@broadinstitute.org"},
]
description = "Parsing and working with ClinVar Variation XML in relational form"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["bioinformatics", "clinvar"]
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "python-dotenv~=1.0.0",
    "xmltodict==0.13.0",
    "coloredlogs",
    "click~=8.1.7",
    "google-cloud-bigquery~=3.13.0",
    "google-cloud-storage~=2.13.0",
    "fastapi~=0.104.1",
    "uvicorn~=0.24.0",
    "requests~=2.31.0"
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "ipykernel",
    "black~=23.9.1",
    "isort~=5.12.0",
    "ruff~=0.1.5",
    "pytest~=7.4.3",
    "pylint~=3.0.2",
    "httpx~=0.25.2"
]

[project.scripts]
clinvar-ingest = "clinvar_ingest.main:main"

[tool.setuptools.packages.find]
include = ["clinvar_ingest*"]

[tool.setuptools.package-data]
"clinvar_ingest" = ["*.json", ".*.env"]
"clinvar_ingest.cloud.bigquery.bq_json_schemas" = ["*.json"]

[tool.isort]
profile = "black"
